Optimize Transport handling
Transports are now skipped when generating constraints and variables.
Instead, one variable representing the flow from one Bus to the next one
is generated and used in the respective Bus balances as appropriate.
This should save us 50% of the variables for each Transport and each
timestep as well as all constraints on a given Transport.